+++ python_modules/libraries/dagster-dg/dagster_dg/cli/env.py
@@ -1,6 +1,8 @@
 from pathlib import Path
+from typing import Optional
 
 import click
+from dagster_shared import check
 from rich.console import Console
 from rich.table import Table
 
@@ -40,3 +42,47 @@ def list_env_command(**global_options: object) -> None:
         table.add_row(key, value)
     console = Console()
     console.print(table)
+
+
+# ########################
+# ##### SET UNSET
+# ########################
+
+
+@env_group.command(name="set", cls=DgClickCommand)
+@click.argument("key", type=str)
+@click.argument("value", type=str, required=False)
+@dg_global_options
+def set_env_command(key: str, value: Optional[str], **global_options: object) -> None:
+    """Set environment variables for the current project."""
+    check.invariant(
+        value is not None or "=" in key,
+        "Input must be of the form `dg set env KEY=VALUE` or `dg set env KEY VALUE`",
+    )
+    if "=" in key:
+        key, value = key.split("=", 1)
+    else:
+        value = None
+
+    cli_config = normalize_cli_config(global_options, click.get_current_context())
+    dg_context = DgContext.for_project_environment(Path.cwd(), cli_config)
+
+    env = Env.from_ctx(dg_context).with_values({key: value})
+    env.write()
+
+
+@env_group.command(name="unset", cls=DgClickCommand)
+@click.argument("key", type=str)
+@dg_global_options
+def unset_env_command(key: str, **global_options: object) -> None:
+    """Unset environment variable for the current project."""
+    cli_config = normalize_cli_config(global_options, click.get_current_context())
+    dg_context = DgContext.for_project_environment(Path.cwd(), cli_config)
+
+    check.invariant(
+        key in Env.from_ctx(dg_context).values,
+        f"Environment variable {key} is not set.",
+    )
+
+    env = Env.from_ctx(dg_context).without_values({key})
+    env.write()
