+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/test_env_commands.py
@@ -43,3 +43,32 @@ def test_list_env_succeeds():
                └─────────┴───────┘
         """).strip()
         )
+
+
+# ########################
+# ##### SET AND UNSET
+# ########################
+
+
+def test_set_env_succeeds():
+    with (
+        ProxyRunner.test(use_fixed_test_components=True) as runner,
+        isolated_example_project_foo_bar(runner, in_workspace=False),
+    ):
+        result = runner.invoke("env", "set", "FOO=bar")
+        assert_runner_result(result)
+        assert Path(".env").read_text() == "FOO=bar"
+
+        result = runner.invoke("env", "set", "FOO=baz")
+        assert_runner_result(result)
+        assert Path(".env").read_text() == "FOO=baz"
+
+        result = runner.invoke("env", "set", "BAR=quux")
+        assert_runner_result(result)
+
+        result = runner.invoke("env", "unset", "FOO")
+        assert_runner_result(result)
+        assert Path(".env").read_text() == "BAR=quux"
+
+        result = runner.invoke("env", "unset", "FOO")
+        assert_runner_result(result, exit_0=False)
